add variable.jsx -store page constant- (to add to Links and Routes)
With Parcel and React-Router, there is a conflict to serve gh-pages: Routes and Links only consider the domain URL (alelubos.github.io) without the repository page (/myFlix-client). Therefore I created a constant to exported and added it to all Route(s) and Link(s) as part of their path (in several components).